Added tests project.

+++ DynamicPriceCore/ViewModels/ProductViewModel.cs
@@ -4,8 +4,8 @@ public class ProductViewModel
 {
 	public int ProductId { get; set; }
 	public string Title { get; set; }
-	public double Price { get; set; }
-	public double MinimumPrice { get; set; }
+	public decimal Price { get; set; }
+	public decimal MinimumPrice { get; set; }
 	public int? Quantity { get; set; }
 	public string? Description { get; set; }
 	public PriceDynamicViewModel[] PriceDynamics { get; set; }
